create a dictionary of methods available in the sciencelab instace
fixes #35
